requirements.txt
flask
openai
openai-whisper
moviepy
ffmpeg-python
slack_sdk
torch